requirements.txt
openai>=1.40.0
gradio>=4.31.0
python-dotenv>=1.0.1
